requirements.txt
fastapi
uvicorn
pandas
psycopg2-binary
openpyxl
pydantic
supabase
